+++ Modules/Intent.Modules.ModuleBuilder.CSharp/Api/NuGetDependencyModel.cs
@@ -11,14 +11,14 @@ using Intent.RoslynWeaver.Attributes;
 namespace Intent.ModuleBuilder.CSharp.Api
 {
     [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
-    public class PackageVersionModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper
+    public class NuGetDependencyModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper
     {
-        public const string SpecializationType = "Package Version";
-        public const string SpecializationTypeId = "231f8cf8-517b-4801-9682-991d22f4e662";
+        public const string SpecializationType = "NuGet Dependency";
+        public const string SpecializationTypeId = "3097322a-a058-4058-beed-4fcd6272f61d";
         protected readonly IElement _element;
 
         [IntentManaged(Mode.Fully)]
-        public PackageVersionModel(IElement element, string requiredType = SpecializationType)
+        public NuGetDependencyModel(IElement element, string requiredType = SpecializationType)
         {
             if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
             {
@@ -35,19 +35,16 @@ namespace Intent.ModuleBuilder.CSharp.Api
 
         public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
 
-        public IElement InternalElement => _element;
+        public string Value => _element.Value;
 
-        public IList<NuGetDependencyModel> NuGetDependencies => _element.ChildElements
-            .GetElementsOfType(NuGetDependencyModel.SpecializationTypeId)
-            .Select(x => new NuGetDependencyModel(x))
-            .ToList();
+        public IElement InternalElement => _element;
 
         public override string ToString()
         {
             return _element.ToString();
         }
 
-        public bool Equals(PackageVersionModel other)
+        public bool Equals(NuGetDependencyModel other)
         {
             return Equals(_element, other?._element);
         }
@@ -57,7 +54,7 @@ namespace Intent.ModuleBuilder.CSharp.Api
             if (ReferenceEquals(null, obj)) return false;
             if (ReferenceEquals(this, obj)) return true;
             if (obj.GetType() != this.GetType()) return false;
-            return Equals((PackageVersionModel)obj);
+            return Equals((NuGetDependencyModel)obj);
         }
 
         public override int GetHashCode()
@@ -67,17 +64,17 @@ namespace Intent.ModuleBuilder.CSharp.Api
     }
 
     [IntentManaged(Mode.Fully)]
-    public static class PackageVersionModelExtensions
+    public static class NuGetDependencyModelExtensions
     {
 
-        public static bool IsPackageVersionModel(this ICanBeReferencedType type)
+        public static bool IsNuGetDependencyModel(this ICanBeReferencedType type)
         {
-            return type != null && type is IElement element && element.SpecializationTypeId == PackageVersionModel.SpecializationTypeId;
+            return type != null && type is IElement element && element.SpecializationTypeId == NuGetDependencyModel.SpecializationTypeId;
         }
 
-        public static PackageVersionModel AsPackageVersionModel(this ICanBeReferencedType type)
+        public static NuGetDependencyModel AsNuGetDependencyModel(this ICanBeReferencedType type)
         {
-            return type.IsPackageVersionModel() ? new PackageVersionModel((IElement)type) : null;
+            return type.IsNuGetDependencyModel() ? new NuGetDependencyModel((IElement)type) : null;
         }
     }
 }