+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -3,6 +3,11 @@
 - Improvement: Added support for Razor file code management.
 - Fixed: Nullable aware assignments of instantiated objects via mapping systems not resolving replacements correctly, leading to uncompilable code.
 
+### Version 3.7.6
+
+- Improvement: `TryGetMaxNetAppVersion` and `GetMaxNetAppVersion` support `.NetStandard2.*` versions.
+- Improvement: Added support for NuGet package dependency modeling.
+
 ### Version 3.7.5
 
 - Fixed: Indentation issue with `CSharpInvocationStatements` through its arguments when attempting to chain them together.