+++ Modules/Intent.Modules.ModuleBuilder.CSharp/release-notes.md
@@ -2,6 +2,10 @@
 
 - Improvement: Added support for generating Razor templates.
 
+### Version 3.6.2
+
+- Improvement: Added support for NuGet Package dependencies.
+
 ### Version 3.6.1
 
 - Improvement: Fixed bug around using generation of NugetPackages template.